requirements.txt
# Core Requirements for Conviction AI Strategy
# These are pinned exact versions for reproducibility

# Core Packages
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tqdm==4.65.0
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.5

# AWS & SageMaker
boto3==1.28.17
sagemaker==2.176.0
awscli==1.27.167
pyarrow==14.0.0

# Configuration
PyYAML==6.0.1

# Machine Learning Libraries
xgboost==1.7.6  # Matches SageMaker container version
lightgbm==4.1.0
catboost==1.2.2
optuna==3.4.0
shap==0.42.1

# Time Series Libraries
statsmodels==0.14.0

# Visualization
plotly==5.18.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
# -------------------------------------------------------------

# Time Series Libraries
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.2.0

# Reinforcement Learning
stable-baselines3[extra]>=2.2.0
gymnasium>=0.29.0

# Financial Libraries - Basic
yfinance>=0.2.36
pandas-ta>=0.3.14b

# Sentiment Analysis & NLP
datasets>=2.16.0
evaluate>=0.4.1
sentence-transformers>=2.3.0

# Visualization & Monitoring
plotly>=5.14.0
tensorboard>=2.13.0
